+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.1.2 ( 9 August 2021)
+
+* Add --exclude-regions option to health command
+* Add --include-regions option to health command
+
 ## v0.1.2
 
 Fix bug for trusted advisor where metadata doesn't come in expected order